refactor(cypress): move partnership tests into subfolder
